requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
dash==3.0.4
dash-bootstrap-components==2.0.2
dnspython==2.7.0
dotenv==0.9.9
Flask==3.0.3
fonttools==4.57.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-secret-manager==2.23.3
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.37.1
nest-asyncio==1.6.0
numpy==2.2.5
openai>=1.0.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
retrying==1.3.4
rsa==4.9.1
setuptools==80.0.1
six==1.17.0
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.21.0
